WIP: enable support for writing to memory

+++ pyogrio/_io.pyx
@@ -6,16 +6,19 @@
 
 import contextlib
 import datetime
+from io import BytesIO
 import locale
 import logging
 import math
 import os
+from pathlib import Path
 import sys
+from uuid import uuid4
 import warnings
 
 from libc.stdint cimport uint8_t, uintptr_t
 from libc.stdlib cimport malloc, free
-from libc.string cimport strlen
+from libc.string cimport memcpy, strlen
 from libc.math cimport isnan
 from cpython.pycapsule cimport PyCapsule_GetPointer
 
@@ -29,7 +32,7 @@ from pyogrio._err cimport *
 from pyogrio._err import CPLE_BaseError, CPLE_NotSupportedError, NullPointerError
 from pyogrio._geometry cimport get_geometry_type, get_geometry_type_code
 from pyogrio.errors import CRSError, DataSourceError, DataLayerError, GeometryError, FieldError, FeatureError
-
+from pyogrio._ogr import _get_driver_metadata_item
 
 log = logging.getLogger(__name__)
 
@@ -173,6 +176,17 @@ cdef const char* override_threadlocal_config_option(str key, str value):
 
 
 cdef void* ogr_open(const char* path_c, int mode, char** options) except NULL:
+    """Open an existing OGR data source
+
+    Parameters
+    ----------
+    path_c : char *
+        input path, including an in-memory path (/vsimem/...)
+    mode : int
+        set to 1 to allow updating data source
+    options : char **, optional
+        dataset open options
+    """
     cdef void* ogr_dataset = NULL
 
     # Force linear approximations in all cases
@@ -1959,10 +1973,20 @@ cdef infer_field_types(list dtypes):
 
 
 cdef create_ogr_dataset_layer(
-    str path, str layer, str driver, str crs, str geometry_type, str encoding,
-    object dataset_kwargs, object layer_kwargs, bint append,
-    dataset_metadata, layer_metadata,
-    OGRDataSourceH* ogr_dataset_out, OGRLayerH* ogr_layer_out,
+    str path,
+    bint is_vsi,
+    str layer,
+    str driver,
+    str crs,
+    str geometry_type,
+    str encoding,
+    object dataset_kwargs,
+    object layer_kwargs,
+    bint append,
+    dataset_metadata,
+    layer_metadata,
+    OGRDataSourceH* ogr_dataset_out,
+    OGRLayerH* ogr_layer_out,
 ):
     """
     Construct the OGRDataSource and OGRLayer objects based on input
@@ -2004,6 +2028,8 @@ cdef create_ogr_dataset_layer(
     cdef OGRSpatialReferenceH ogr_crs = NULL
     cdef OGRwkbGeometryType geometry_code
     cdef int layer_idx = -1
+    cdef VSIStatBufL st_buf
+
 
     path_b = path.encode('UTF-8')
     path_c = path_b
@@ -2011,18 +2037,29 @@ cdef create_ogr_dataset_layer(
     driver_b = driver.encode('UTF-8')
     driver_c = driver_b
 
+    if is_vsi:
+        path_exists = VSIStatExL(path_c, &st_buf, VSI_STAT_EXISTS_FLAG) == 0
+    else:
+        path_exists = os.path.exists(path)
+
     if not layer:
         layer = os.path.splitext(os.path.split(path)[1])[0]
 
     # if shapefile, GeoJSON, or FlatGeobuf, always delete first
     # for other types, check if we can create layers
     # GPKG might be the only multi-layer writeable type.  TODO: check this
-    if driver in ('ESRI Shapefile', 'GeoJSON', 'GeoJSONSeq', 'FlatGeobuf') and os.path.exists(path):
+    if driver in ('ESRI Shapefile', 'GeoJSON', 'GeoJSONSeq', 'FlatGeobuf') and path_exists:
         if not append:
-            os.unlink(path)
+            if is_vsi:
+                VSIUnlink(path_c)
+            else:
+                os.unlink(path)
+
+            path_exists = False
+
 
     layer_exists = False
-    if os.path.exists(path):
+    if path_exists:
         try:
             ogr_dataset = ogr_open(path_c, 1, NULL)
 
@@ -2044,7 +2081,11 @@ cdef create_ogr_dataset_layer(
                 raise exc
 
             # otherwise create from scratch
-            os.unlink(path)
+            if is_vsi:
+                VSIUnlink(path_c)
+            else:
+                os.unlink(path)
+
             ogr_dataset = NULL
 
     # either it didn't exist or could not open it in write mode
@@ -2133,15 +2174,29 @@ cdef create_ogr_dataset_layer(
 
     ogr_dataset_out[0] = ogr_dataset
     ogr_layer_out[0] = ogr_layer
+
     return create_layer
 
 
 # TODO: set geometry and field data as memory views?
 def ogr_write(
-    str path, str layer, str driver, geometry, fields, field_data, field_mask,
-    str crs, str geometry_type, str encoding, object dataset_kwargs,
-    object layer_kwargs, bint promote_to_multi=False, bint nan_as_null=True,
-    bint append=False, dataset_metadata=None, layer_metadata=None,
+    object path_or_fp,
+    str layer,
+    str driver,
+    geometry,
+    fields,
+    field_data,
+    field_mask,
+    str crs,
+    str geometry_type,
+    str encoding,
+    object dataset_kwargs,
+    object layer_kwargs,
+    bint promote_to_multi=False,
+    bint nan_as_null=True,
+    bint append=False,
+    dataset_metadata=None,
+    layer_metadata=None,
     gdal_tz_offsets=None
 ):
     cdef OGRDataSourceH ogr_dataset = NULL
@@ -2158,6 +2213,10 @@ def ogr_write(
     cdef int num_records = -1
     cdef int num_field_data = len(field_data) if field_data is not None else 0
     cdef int num_fields = len(fields) if fields is not None else 0
+    cdef VSILFILE *vsi_handle = NULL
+    cdef bytes vsi_bytes_buffer
+    cdef unsigned char *vsi_buffer = NULL
+    cdef vsi_l_offset vsi_buffer_size = 0
 
     if num_fields != num_field_data:
         raise ValueError("field_data array needs to be same length as fields array")
@@ -2196,9 +2255,54 @@ def ogr_write(
     if gdal_tz_offsets is None:
         gdal_tz_offsets = {}
 
+
+    ### Setup in-memory handler if needed
+    if isinstance(path_or_fp, BytesIO):
+        # TODO: add try / catch around this and cleanup open handles
+
+        # Create in-memory directory to contain auxiliary files
+        memfilename = uuid4().hex
+        VSIMkdir(f"/vsimem/{memfilename}".encode("utf-8"), 0666)
+
+        # file extension is required for some drivers, set it based on driver metadata
+        ext = ''
+        recommended_ext = _get_driver_metadata_item(driver, "DMD_EXTENSIONS")
+        if recommended_ext is not None:
+            ext = "." + recommended_ext.split(' ')[0]
+
+        path = f"/vsimem/{memfilename}/{memfilename}{ext}"
+
+        # TODO: automatically zip shapefile or raise Exception?
+        # Otherwise no way to get back from memory in useful way
+        # if driver == "ESRI Shapefile":
+        #     path += ".zip"
+
+        # copy existing bytes into VSI file
+        path_or_fp.seek(0)
+        vsi_bytes_buffer = path_or_fp.read()
+        vsi_buffer_size = len(vsi_bytes_buffer)
+
+        if vsi_buffer_size > 0:
+            # copy and transfer memory to GDAL
+            vsi_buffer = <unsigned char *>CPLMalloc(vsi_buffer_size)
+            memcpy(<void *>vsi_bytes_buffer, vsi_buffer, vsi_buffer_size)
+            # vsi_handle = VSIFileFromMemBuffer(path.encode('UTF-8'), <unsigned char *>vsi_bytes_buffer, vsi_buffer_size, 0)
+            vsi_handle = VSIFileFromMemBuffer(path.encode('UTF-8'), <unsigned char *>vsi_buffer, vsi_buffer_size, 1)
+            if vsi_handle == NULL:
+                raise RuntimeError("in-memory file could not be created")
+
+            vsi_buffer = NULL # to prevent cleanup later (GDAL will free)
+
+        else:
+            vsi_handle = VSIFOpenL(path.encode('UTF-8'), "w")
+
+    else:
+        path = path_or_fp
+
+
     ### Setup up dataset and layer
     layer_created = create_ogr_dataset_layer(
-        path, layer, driver, crs, geometry_type, encoding,
+        path, vsi_handle!=NULL, layer, driver, crs, geometry_type, encoding,
         dataset_kwargs, layer_kwargs, append,
         dataset_metadata, layer_metadata,
         &ogr_dataset, &ogr_layer,
@@ -2413,12 +2517,45 @@ def ogr_write(
     ### Final cleanup
     if ogr_dataset != NULL:
         GDALClose(ogr_dataset)
+        ogr_dataset = NULL
 
         # GDAL will set an error if there was an error writing the data source
         # on close
         exc = exc_check()
         if exc:
             raise DataSourceError(f"Failed to write features to dataset {path}; {exc}")
+
+
+    # TODO: merge this in with updated try / finally blocks from #396
+    if vsi_handle != NULL:
+        try:
+            # rewind to beginning of file and read bytes back from memory
+            VSIFSeekL(vsi_handle, 0, 0)
+
+            # Take ownership of the buffer to avoid a copy; GDAL will automatically
+            # unlink the memory file
+            vsi_buffer = VSIGetMemFileBuffer(path.encode("UTF-8"), &vsi_buffer_size, 1)
+            if vsi_buffer == NULL:
+                # TODO: make sure following cleanup steps are called before final return
+                raise RuntimeError("could not read bytes from in-memory file")
+
+            # reset path_or_fp and write bytes to it
+            path_or_fp.seek(0)
+            path_or_fp.truncate()  # truncate in case bytes were passed in (e.g., append / add layer)
+            path_or_fp.write(<bytes>vsi_buffer[:vsi_buffer_size])
+            # rewind to beginning to allow caller to read
+            path_or_fp.seek(0)
+
+        finally:
+            # TODO: co this in the outer finally block after #396
+            CPLFree(vsi_buffer)
+
+            if vsi_handle != NULL:
+                VSIFCloseL(vsi_handle)
+
+            # remove virtual directory created above and unlink all files
+            VSIRmdirRecursive(str(Path(path).parent).encode("UTF-8"))
+            vsi_handle = NULL
 
 
 def ogr_write_arrow(
@@ -2442,6 +2579,7 @@ def ogr_write_arrow(
     cdef OGRDataSourceH ogr_dataset = NULL
     cdef OGRLayerH ogr_layer = NULL
     cdef char **options = NULL
+    cdef VSILFILE *vsi_handle = NULL
     cdef ArrowArrayStream* stream = NULL
     cdef ArrowSchema schema
     cdef ArrowArray array
@@ -2450,7 +2588,7 @@ def ogr_write_arrow(
     array.release = NULL
 
     layer_created = create_ogr_dataset_layer(
-        path, layer, driver, crs, geometry_type, encoding,
+        path, vsi_handle!=NULL, layer, driver, crs, geometry_type, encoding,
         dataset_kwargs, layer_kwargs, append,
         dataset_metadata, layer_metadata,
         &ogr_dataset, &ogr_layer,
